+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-23 15:50:44
-// Commit: style(test): adjust design tokens, typography, and contrast in test
+// Updated at: 2026-05-24 13:28:33
+// Commit: feat(test): integrate custom telemetry logger in test
 
 class testComponentState {
-  final String commitRef = '1351';
-  final String lastMessage = 'style(test): adjust design tokens, typography, and contrast in test';
-  final DateTime timestamp = DateTime.parse('2026-05-23T15:50:44+05:00');
+  final String commitRef = '1388';
+  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
+  final DateTime timestamp = DateTime.parse('2026-05-24T13:28:33+05:00');
 }